rename ui.features to ui.flows
