--- app/src/main/java/com/methodsignature/healthyrecipes/ui/features/content/ContentFlow.kt
@@ -1,12 +1,0 @@
-package com.methodsignature.healthyrecipes.ui.features.content
-
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-
-@Composable
-fun ContentFlow(
-    modifier: Modifier = Modifier,
-) {
-    Text(text = "Content Flow")
-}
