+++ app/src/test/java/com/methodsignature/healthyrecipes/ui/flows/onboarding/OnboardingFlowViewModelTest.kt
@@ -1,8 +1,8 @@
-package com.methodsignature.healthyrecipes.ui.features.onboarding
+package com.methodsignature.healthyrecipes.ui.flows.onboarding
 
 import app.cash.turbine.test
 import com.methodsignature.healthyrecipes.BaseTest
-import com.methodsignature.healthyrecipes.ui.features.onboarding.OnboardingFlowViewModel.NavigationEvent
+import com.methodsignature.healthyrecipes.ui.flows.onboarding.OnboardingFlowViewModel.NavigationEvent
 import kotlinx.coroutines.test.runTest
 import org.amshove.kluent.shouldBeEqualTo
 import kotlin.test.Test
